+ Extract common config validation checks (config_validate lib) + Remove redundant checks of duplicate Events and Fields (CheckContainsEvent, CheckContainsField) + Change ParseLolaxxx functions to return void instead of ignoring the return
