+++ score/mw/com/impl/configuration/config_parser.cpp
@@ -11,6 +11,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 #include "score/mw/com/impl/configuration/config_parser.h"
+#include "score/mw/com/impl/configuration/config_validate.h"
 
 #include "score/mw/com/impl/configuration/configuration_common_resources.h"
 #include "score/mw/com/impl/configuration/lola_method_instance_deployment.h"
@@ -30,7 +31,6 @@
 #include <cstdlib>
 #include <exception>
 #include <optional>
-#include <set>
 #include <string>
 #include <string_view>
 #include <tuple>
@@ -137,13 +137,7 @@ auto ParseInstanceSpecifier(const score::json::Object& json_map) -> InstanceSpec
     SCORE_LANGUAGE_FUTURECPP_PRECONDITION_PRD_MESSAGE(string_result.has_value(),
                                                       "Configuration corrupted, check with json schema");
     auto instance_specifier_name = string_result.value().get();
-    const auto instance_specifier_result = InstanceSpecifier::Create(std::move(instance_specifier_name));
-    if (!instance_specifier_result.has_value())
-    {
-        score::mw::log::LogFatal("lola") << "Invalid InstanceSpecifier.";
-        SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-    }
-    return instance_specifier_result.value();
+    return CreateValidInstanceSpecifier(std::move(instance_specifier_name));
 }
 
 auto ParseServiceTypeName(const score::json::Object& json_map) -> const std::string&
@@ -331,28 +325,6 @@ class ServiceElementInstanceDeploymentParser
         return name_value.value().get();
     }
 
-    void CheckContainsEvent(const score::json::Object::const_iterator name,
-                            const LolaServiceInstanceDeployment& service)
-    {
-        const auto name_value = GetName(name);
-        if (service.ContainsEvent(name_value))
-        {
-            score::mw::log::LogFatal("lola") << "Event Name Duplicated. Not allowed";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
-    }
-
-    void CheckContainsField(const score::json::Object::const_iterator name,
-                            const LolaServiceInstanceDeployment& service)
-    {
-        const auto name_value = GetName(name);
-        if (service.ContainsField(name_value))
-        {
-            score::mw::log::LogFatal("lola") << "Field Name Duplicated. Not allowed";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
-    }
-
     template <typename element_type>
     std::optional<element_type> RetrieveJsonElement(const score::json::Object::const_iterator element_iterator)
     {
@@ -424,8 +396,6 @@ auto ParseLolaEventInstanceDeployment(const score::json::Object& json_map, LolaS
         ServiceElementInstanceDeploymentParser deployment_parser{event_object};
 
         const auto& event_name_it = event_object.find(kEventNameKey.data());
-        deployment_parser.CheckContainsEvent(event_name_it, service);
-
         auto event_name_value = deployment_parser.GetName(event_name_it);
 
         const auto number_of_sample_slots = deployment_parser.GetNumberOfSampleSlots();
@@ -446,10 +416,7 @@ auto ParseLolaEventInstanceDeployment(const score::json::Object& json_map, LolaS
                                                             enforce_max_samples,
                                                             number_of_tracing_slots);
 
-        const auto emplace_result = service.events_.emplace(std::piecewise_construct,
-                                                            std::forward_as_tuple(std::move(event_name_value)),
-                                                            std::forward_as_tuple(event_deployment));
-        SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD_MESSAGE(emplace_result.second, "Could not emplace element in map");
+        EmplaceOrFatal(service.events_, std::move(event_name_value), event_deployment, "An event instance");
     }
 }
 
@@ -480,8 +447,6 @@ auto ParseLolaFieldInstanceDeployment(const score::json::Object& json_map, LolaS
         ServiceElementInstanceDeploymentParser deployment_parser{field_object};
 
         const auto& field_name_it = field_object.find(kFieldNameKey);
-        deployment_parser.CheckContainsField(field_name_it, service);
-
         auto field_name_value = deployment_parser.GetName(field_name_it);
 
         const auto number_of_sample_slots =
@@ -508,10 +473,7 @@ auto ParseLolaFieldInstanceDeployment(const score::json::Object& json_map, LolaS
                                                                     number_of_tracing_slots),
                                         use_get_if_available,
                                         use_set_if_available);
-        const auto emplace_result = service.fields_.emplace(std::piecewise_construct,
-                                                            std::forward_as_tuple(std::move(field_name_value)),
-                                                            std::forward_as_tuple(field_deployment));
-        SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD_MESSAGE(emplace_result.second, "Could not emplace element in map");
+        EmplaceOrFatal(service.fields_, std::move(field_name_value), field_deployment, "A field instance");
     }
 }
 
@@ -543,9 +505,7 @@ auto ParseLolaMethodInstanceDeployment(const score::json::Object& json_map, Lola
             GetOptionalValueFromJson<bool>(method_object, kMethodEnabledKey).value_or(kMethodEnabledDefaultValue);
         const LolaMethodInstanceDeployment method_deployment{queue_size, method_enabled};
 
-        const auto emplace_result = service.methods_.emplace(
-            std::piecewise_construct, std::forward_as_tuple(method_name), std::forward_as_tuple(method_deployment));
-        SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD_MESSAGE(emplace_result.second, "Could not emplace element in map");
+        EmplaceOrFatal(service.methods_, method_name, method_deployment, "A method instance");
     }
 }
 
@@ -798,33 +758,24 @@ auto ParseServiceInstances(const score::json::Object& object, TracingConfigurati
 
         auto instance_deployments = ParseServiceInstanceDeployments(
             service_instance_map, tracing_configuration, service_identifier, instanceSpecifier);
-        if (instance_deployments.size() != 1U)
-        {
-            score::mw::log::LogFatal("lola") << "More or less then one deployment for " << service_identifier.ToString()
-                                             << ". Multi-Binding support right now not supported";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
+        ValidateSingleDeployment(instance_deployments, service_identifier);
 
-        auto emplaceRes = service_instance_deployments.emplace(std::piecewise_construct,
-                                                               std::forward_as_tuple(instanceSpecifier),
-                                                               std::forward_as_tuple(instance_deployments.at(0U)));
-        if (emplaceRes.second == false)
-        {
-            score::mw::log::LogFatal("lola") << "Unexpected error, when inserting service instance deployments.";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
+        EmplaceOrFatal(service_instance_deployments,
+                       instanceSpecifier,
+                       instance_deployments.at(0U),
+                       "Service instance deployment");
     }
     return service_instance_deployments;
 }
 
 // See Note 1
 // coverity[autosar_cpp14_a15_5_3_violation]
-auto ParseLolaEventTypeDeployments(const score::json::Object& json_map, LolaServiceTypeDeployment& service) -> bool
+void ParseLolaEventTypeDeployments(const score::json::Object& json_map, LolaServiceTypeDeployment& service)
 {
     const auto& events = json_map.find(kEventsKey.data());
     if (events == json_map.cend())
     {
-        return false;
+        return;
     }
     auto events_list_result = events->second.As<score::json::List>();
     SCORE_LANGUAGE_FUTURECPP_PRECONDITION_PRD_MESSAGE(events_list_result.has_value(),
@@ -849,27 +800,18 @@ auto ParseLolaEventTypeDeployments(const score::json::Object& json_map, LolaServ
         const auto event_id_casted = event_id->second.As<std::uint16_t>();
         SCORE_LANGUAGE_FUTURECPP_PRECONDITION_PRD_MESSAGE(event_id_casted.has_value(),
                                                           "Configuration corrupted, check with json schema");
-        const auto result = service.events_.emplace(std::piecewise_construct,
-                                                    std::forward_as_tuple(event_name_casted.value().get()),
-                                                    std::forward_as_tuple(event_id_casted.value()));
-
-        if (result.second != true)
-        {
-            score::mw::log::LogFatal("lola") << "An event was configured twice.";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
+        EmplaceOrFatal(service.events_, event_name_casted.value().get(), event_id_casted.value(), "An event");
     }
-    return true;
 }
 
 // See Note 1
 // coverity[autosar_cpp14_a15_5_3_violation]
-auto ParseLolaFieldTypeDeployments(const score::json::Object& json_map, LolaServiceTypeDeployment& service) -> bool
+void ParseLolaFieldTypeDeployments(const score::json::Object& json_map, LolaServiceTypeDeployment& service)
 {
     const auto& fields = json_map.find(kFieldsKey.data());
     if (fields == json_map.cend())
     {
-        return false;
+        return;
     }
 
     auto fields_list_result = fields->second.As<score::json::List>();
@@ -895,27 +837,18 @@ auto ParseLolaFieldTypeDeployments(const score::json::Object& json_map, LolaServ
         const auto field_id_casted = field_id->second.As<std::uint16_t>();
         SCORE_LANGUAGE_FUTURECPP_PRECONDITION_PRD_MESSAGE(field_id_casted.has_value(),
                                                           "Configuration corrupted, check with json schema");
-        const auto result = service.fields_.emplace(std::piecewise_construct,
-                                                    std::forward_as_tuple(field_name_casted.value().get()),
-                                                    std::forward_as_tuple(field_id_casted.value()));
-
-        if (result.second != true)
-        {
-            score::mw::log::LogFatal("lola") << "A field was configured twice.";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
+        EmplaceOrFatal(service.fields_, field_name_casted.value().get(), field_id_casted.value(), "A field");
     }
-    return true;
 }
 
 // See Note 1
 // coverity[autosar_cpp14_a15_5_3_violation]
-auto ParseLolaMethodTypeDeployments(const score::json::Object& json_map, LolaServiceTypeDeployment& service) -> bool
+void ParseLolaMethodTypeDeployments(const score::json::Object& json_map, LolaServiceTypeDeployment& service)
 {
     const auto& methods = json_map.find(kMethodsKey.data());
     if (methods == json_map.cend())
     {
-        return false;
+        return;
     }
 
     auto methods_list_result = methods->second.As<score::json::List>();
@@ -941,58 +874,8 @@ auto ParseLolaMethodTypeDeployments(const score::json::Object& json_map, LolaSer
         const auto method_id_casted = method_id->second.As<std::uint16_t>();
         SCORE_LANGUAGE_FUTURECPP_PRECONDITION_PRD_MESSAGE(method_id_casted.has_value(),
                                                           "Configuration corrupted, check with json schema");
-        const auto result = service.methods_.emplace(std::piecewise_construct,
-                                                     std::forward_as_tuple(method_name_casted.value().get()),
-                                                     std::forward_as_tuple(method_id_casted.value()));
-
-        if (result.second != true)
-        {
-            score::mw::log::LogFatal("lola") << "A method was configured twice.";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
+        EmplaceOrFatal(service.methods_, method_name_casted.value().get(), method_id_casted.value(), "A method");
     }
-    return true;
-}
-
-auto AreEventFieldAndMethodIdsUnique(const LolaServiceTypeDeployment& lola_service_type_deployment) -> bool
-{
-    const auto& events = lola_service_type_deployment.events_;
-    const auto& fields = lola_service_type_deployment.fields_;
-    const auto& methods = lola_service_type_deployment.methods_;
-
-    static_assert(std::is_same<LolaEventId, LolaFieldId>::value,
-                  "EventId and FieldId should have the same underlying type.");
-    static_assert(std::is_same<LolaEventId, LolaMethodId>::value,
-                  "EventId and MethodId should have the same underlying type.");
-    std::set<LolaEventId> ids{};
-
-    for (const auto& event : events)
-    {
-        const auto result = ids.insert(event.second);
-        if (!result.second)
-        {
-            return false;
-        }
-    }
-
-    for (const auto& field : fields)
-    {
-        const auto result = ids.insert(field.second);
-        if (!result.second)
-        {
-            return false;
-        }
-    }
-
-    for (const auto& method : methods)
-    {
-        const auto result = ids.insert(method.second);
-        if (!result.second)
-        {
-            return false;
-        }
-    }
-    return true;
 }
 
 // See Note 1
@@ -1007,19 +890,10 @@ auto ParseLoLaServiceTypeDeployments(const score::json::Object& json_map) -> Lol
     SCORE_LANGUAGE_FUTURECPP_PRECONDITION_PRD_MESSAGE(service_id_casted.has_value(),
                                                       "Configuration corrupted, check with json schema");
     LolaServiceTypeDeployment lola{service_id_casted.value()};
-    const bool events_exist = ParseLolaEventTypeDeployments(json_map, lola);
-    const bool fields_exist = ParseLolaFieldTypeDeployments(json_map, lola);
-    const bool methods_exist = ParseLolaMethodTypeDeployments(json_map, lola);
-    if (!events_exist && !fields_exist && !methods_exist)
-    {
-        score::mw::log::LogFatal("lola") << "Configuration should contain at least one event, field, or method.";
-        SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-    }
-    if (!AreEventFieldAndMethodIdsUnique(lola))
-    {
-        score::mw::log::LogFatal("lola") << "Configuration cannot contain duplicate eventId, fieldId, or methodId.";
-        SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-    }
+    ParseLolaEventTypeDeployments(json_map, lola);
+    ParseLolaFieldTypeDeployments(json_map, lola);
+    ParseLolaMethodTypeDeployments(json_map, lola);
+    ValidateUniqueServiceElementIds(lola);
     return lola;
 }
 
@@ -1085,15 +959,7 @@ auto ParseServiceTypes(const score::json::Object& json_map) -> Configuration::Se
         const auto service_identifier = ParseServiceTypeIdentifier(service_type_map);
 
         const auto service_deployment = ParseServiceTypeDeployment(service_type_map);
-        const auto inserted = service_type_deployments.emplace(std::piecewise_construct,
-                                                               std::forward_as_tuple(service_identifier),
-                                                               std::forward_as_tuple(service_deployment));
-
-        if (inserted.second != true)
-        {
-            score::mw::log::LogFatal("lola") << "Service Type was deployed twice";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
+        EmplaceOrFatal(service_type_deployments, service_identifier, service_deployment, "Service Type");
     }
     return service_type_deployments;
 }
@@ -1319,97 +1185,6 @@ auto ParseTracingProperties(const score::json::Object& top_level_object) -> Trac
             std::string{tracing_filter_config_path.data(), tracing_filter_config_path.size()});
     }
     return tracing_configuration;
-}
-
-void CrosscheckAsilLevels(const Configuration& config)
-{
-    for (const auto& service_instance : config.GetServiceInstances())
-    {
-        if ((service_instance.second.asilLevel_ == QualityType::kASIL_B) &&
-            (config.GetGlobalConfiguration().GetProcessAsilLevel() != QualityType::kASIL_B))
-        {
-            ::score::mw::log::LogFatal("lola")
-                << "Service instance has a higher ASIL than the process. This is invalid, terminating";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
-    }
-}
-
-/**
- * \brief Checks, whether for all (binding) types used in service instances, there is also a corresponding type
- *        in service types.
- * @param config configuration to be crosschecked.
- */
-void CrosscheckServiceInstancesToTypes(const Configuration& config)
-{
-    for (const auto& service_instance : config.GetServiceInstances())
-    {
-        const auto foundServiceType = config.GetServiceTypes().find(service_instance.second.service_);
-        if (foundServiceType == config.GetServiceTypes().cend())
-        {
-            ::score::mw::log::LogFatal("lola")
-                << "Service instance " << service_instance.first << "refers to a service type ("
-                << service_instance.second.service_.ToString()
-                << "), which is not configured. This is invalid, terminating";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
-        // check, that binding in service type and service instance are equal. Since currently ServiceTypeDeployment
-        // only supports LolaServiceTypeDeployment, everything else than LolaServiceInstanceDeployment is an error.
-        // LCOV_EXCL_BR_START: Defensive programming: Parse() currently terminates if the ServiceInstanceDeployment
-        // contains anything other than a Lola binding. Therefore, it's impossible to reach this point without
-        // a LolaServiceInstanceDeployment.
-        if (!std::holds_alternative<LolaServiceInstanceDeployment>(service_instance.second.bindingInfo_))
-        {
-            // LCOV_EXCL_BR_STOP
-            // LCOV_EXCL_START defensive programming: Parse() currently terminates if the ServiceInstanceDeployment
-            // contains anything other than a Lola binding. Therefore, it's impossible to reach this point without
-            // a LolaServiceInstanceDeployment.
-            ::score::mw::log::LogFatal("lola")
-                << "Service instance " << service_instance.first
-                << "refers to an not yet supported binding. This is invalid, terminating";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-            // LCOV_EXCL_STOP
-        }
-        if (!std::holds_alternative<LolaServiceTypeDeployment>(foundServiceType->second.binding_info_))
-        {
-            ::score::mw::log::LogFatal("lola")
-                << "Service type " << service_instance.second.service_.ToString()
-                << "refers to an not yet supported binding. This is invalid, terminating";
-            SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-        }
-        // check, that for each service-element-name in the instance deployment, there exists a corresponding
-        // service-element-name in the type deployment
-        const auto& serviceInstanceDeployment =
-            std::get<LolaServiceInstanceDeployment>(service_instance.second.bindingInfo_);
-        for (const auto& eventInstanceElement : serviceInstanceDeployment.events_)
-        {
-            const auto& serviceTypeDeployment =
-                std::get<LolaServiceTypeDeployment>(foundServiceType->second.binding_info_);
-            const auto search = serviceTypeDeployment.events_.find(eventInstanceElement.first);
-            if (search == serviceTypeDeployment.events_.cend())
-            {
-                ::score::mw::log::LogFatal("lola")
-                    << "Service instance " << service_instance.first << "event" << eventInstanceElement.first
-                    << "refers to an event, which doesn't exist in the referenced service type ("
-                    << service_instance.second.service_.ToString() << "). This is invalid, terminating";
-                SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-            }
-        }
-        for (const auto& fieldInstanceElement : serviceInstanceDeployment.fields_)
-        {
-            const auto& serviceTypeDeployment =
-                std::get<LolaServiceTypeDeployment>(foundServiceType->second.binding_info_);
-            const auto search = serviceTypeDeployment.fields_.find(fieldInstanceElement.first);
-            if (search == serviceTypeDeployment.fields_.cend())
-            {
-                ::score::mw::log::LogFatal("lola")
-                    << "Service instance " << service_instance.first << "field" << fieldInstanceElement.first
-                    << "refers to a field, which doesn't exist in the referenced service type ("
-                    << service_instance.second.service_.ToString() << "). This is invalid, terminating";
-                SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD(false);
-            }
-        }
-    }
 }
 
 }  // namespace
